# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Shared Assets APIs",
        "version": "1.0.0"
    },
    "paths": {
        "/api/assets": {
            "post": {
                "tags": [
                    "Assets"
                ],
                "summary": "Create asset",
                "operationId": "009a09a3ccd2d061f0a7f2e7b40696a9",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "label": {
                                                "type": "string"
                                            },
                                            "content": {
                                                "type": "string"
                                            },
                                            "tag_id": {
                                                "type": "integer"
                                            },
                                            "team_id": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "label": "Price how much",
                                    "content": "some content goes here....",
                                    "tag_id": 1,
                                    "team_id": 1
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "label": {
                                            "type": "string",
                                            "example": "Price how much"
                                        },
                                        "content": {
                                            "type": "string",
                                            "example": "some content goes here...."
                                        },
                                        "tag_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "team_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2022-10-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2022-10-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/assets/update/{id}": {
            "put": {
                "tags": [
                    "Assets"
                ],
                "summary": "Update Asset",
                "operationId": "56e54a1f566572e61441755abd286809",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "label": {
                                                "type": "string"
                                            },
                                            "content": {
                                                "type": "string"
                                            },
                                            "tag_id": {
                                                "type": "integer"
                                            },
                                            "team_id": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "title": "Updated Label",
                                    "content": "Updated content goes here...",
                                    "tag_id": 1,
                                    "team_id": 1
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "label": {
                                            "type": "string",
                                            "example": "Updated Label"
                                        },
                                        "content": {
                                            "type": "string",
                                            "example": "Updated content goes here..."
                                        },
                                        "tag_id": {
                                            "type": "string",
                                            "example": 1
                                        },
                                        "team_id": {
                                            "type": "string",
                                            "example": 1
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/assets/delete/{id}": {
            "delete": {
                "tags": [
                    "Assets"
                ],
                "summary": "Delete Asset",
                "operationId": "0f89dd203bfe1d4dc13b9f32f2cc111f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Asset is deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}